Shared Pointer WindowResCalc

+++ TonicEngine/Sources/General/WindowResolutionCalculator.cpp
@@ -10,19 +10,19 @@ void WindowResolutionCalculator::setResolution(int width, int height) {
 }
 
 int WindowResolutionCalculator::getConvertedHeight(int height) {
-    return height * (_windowHeight / DEFAULT_HEIGHT) + 1;
+    return static_cast<int>(height * (_windowHeight / DEFAULT_HEIGHT) + 1);
 }
 
 int WindowResolutionCalculator::getConvertedWidth(int width) {
-    return width * (_windowWidth / DEFAULT_WIDTH) + 1;
+    return static_cast<int>(width * (_windowWidth / DEFAULT_WIDTH) + 1);
 
 }
 
 int WindowResolutionCalculator::getConvertedxPos(int xPos) {
-    return xPos * (_windowWidth / DEFAULT_WIDTH);
+    return static_cast<int>(xPos * (_windowWidth / DEFAULT_WIDTH));
 }
 
 int WindowResolutionCalculator::getConvertedyPos(int yPos) {
-    return yPos * (_windowHeight / DEFAULT_HEIGHT);
+    return static_cast<int>(yPos * (_windowHeight / DEFAULT_HEIGHT));
 
 }
